# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Larpsec — анализ организационных изменений",
    "description": "Комплекты до/после, функции, потенциальные отклонения и проверяемые источники.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/comparisons": {
      "post": {
        "tags": [
          "comparisons"
        ],
        "summary": "Create Comparison",
        "operationId": "create_comparison_api_v1_comparisons_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComparison"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "comparisons"
        ],
        "summary": "List Comparisons",
        "operationId": "list_comparisons_api_v1_comparisons_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comparison"
                  },
                  "title": "Response List Comparisons Api V1 Comparisons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}": {
      "get": {
        "tags": [
          "comparisons"
        ],
        "summary": "Get Comparison",
        "operationId": "get_comparison_api_v1_comparisons__comparison_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "comparisons"
        ],
        "summary": "Remove Comparison",
        "operationId": "remove_comparison_api_v1_comparisons__comparison_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/coverage": {
      "patch": {
        "tags": [
          "comparisons"
        ],
        "summary": "Update Coverage",
        "operationId": "update_coverage_api_v1_comparisons__comparison_id__coverage_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_comparisons__comparison_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "side",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "before",
                "after"
              ],
              "type": "string",
              "title": "Side"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_comparisons__comparison_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/documents/{document_id}/fragments": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Fragments",
        "operationId": "get_fragments_api_v1_comparisons__comparison_id__documents__document_id__fragments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fragment"
                  },
                  "title": "Response Get Fragments Api V1 Comparisons  Comparison Id  Documents  Document Id  Fragments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/sources/{fragment_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Source",
        "operationId": "get_source_api_v1_comparisons__comparison_id__sources__fragment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "fragment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fragment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fragment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/documents/{document_id}/download": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Download Document",
        "operationId": "download_document_api_v1_comparisons__comparison_id__documents__document_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Remove Document",
        "operationId": "remove_document_api_v1_comparisons__comparison_id__documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Start Analysis",
        "operationId": "start_analysis_api_v1_comparisons__comparison_id__analyze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/result": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Result",
        "operationId": "get_result_api_v1_comparisons__comparison_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons/{comparison_id}/findings/{finding_id}/review": {
      "patch": {
        "tags": [
          "analysis"
        ],
        "summary": "Review Finding",
        "operationId": "review_finding_api_v1_comparisons__comparison_id__findings__finding_id__review_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_comparisons__comparison_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Один DOCX, PDF, XLSX или TXT"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_comparisons__comparison_id__documents_post"
      },
...
      "CoverageUpdate": {
        "properties": {
          "before_complete": {
            "type": "boolean",
            "title": "Before Complete"
          },
          "after_complete": {
            "type": "boolean",
            "title": "After Complete"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "before_complete",
          "after_complete"
        ],
        "title": "CoverageUpdate"
      },
      "CreateComparison": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "default": "Анализ реорганизации"
          },
          "before_complete": {
            "type": "boolean",
            "title": "Before Complete",
            "default": false
          },
          "after_complete": {
            "type": "boolean",
            "title": "After Complete",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateComparison"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Review": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed",
              "dismissed"
            ],
            "title": "Status"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "title": "Comment",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "Review"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}